fixing issues with mutex locks

+++ android/src/main/cpp/native_buffer_api.h
@@ -16,9 +16,9 @@ extern "C" {
 #endif
 
 FFI_PLUGIN_EXPORT int initNativeBufferFFI(const char* key, int capacity, int maxBufferSize);
-FFI_PLUGIN_EXPORT uintptr_t pushVideoNativeBufferFFI(const char* key, const uint8_t* buffer, size_t dataSize,
+FFI_PLUGIN_EXPORT int pushVideoNativeBufferFFI(const char* key, const uint8_t* buffer, size_t dataSize,
   int width, int height, uint64_t frameTime, int rotation, int frameType);
-FFI_PLUGIN_EXPORT uintptr_t pushAudioNativeBufferFFI(const char* key, const uint8_t* buffer, size_t dataSize,
+FFI_PLUGIN_EXPORT int pushAudioNativeBufferFFI(const char* key, const uint8_t* buffer, size_t dataSize,
   int sampleRate, int channels, uint64_t frameTime);
 FFI_PLUGIN_EXPORT uintptr_t popNativeBufferFFI(const char* key);
 FFI_PLUGIN_EXPORT void freeNativeBufferFFI(const char* key);
@@ -27,7 +27,7 @@ FFI_PLUGIN_EXPORT bool initializeDartApiDL(void* data);
 FFI_PLUGIN_EXPORT bool registerDartPort(const char* channel_name, int64_t port);
 
 #ifdef __cplusplus
-}
+} // extern "C"
 #endif
 
 #endif // NATIVE_BUFFER_API_H
